board테이블,localcate테이블 dto,레파지토리,서비스 작성

+++ Life_Fit/src/main/java/org/zerock/life_fit/board/domain/LocalCate.java
@@ -15,7 +15,7 @@ public class LocalCate {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "localnum", updatable = false)
-    private int localnum;
+    private Long localnum;
     @Column(name = "localname", nullable = false)
     private String localname;
 }
